inserting check that mlcube/be_mlcube.yaml (same for ms) exists otherwise exiting so user knows to populate it.

+++ examples/fl_post/fl/be_intel_debug_build.sh
@@ -1,5 +1,10 @@
-cp mlcube/be_mlcube.yaml mlcube/mlcube.yaml
+if [ -f "mlcube/be_mlcube.yaml" ]; then
 
+    cp mlcube/be_mlcube.yaml mlcube/mlcube.yaml
+else
+    echo "mlcube/be_mlcube.yaml was not able to be copied because it did not exist"
+    exit
+fi
 
 while getopts b flag; do
     case "${flag}" in
